Moving pytest fixtures to conftest.py
Now flake8 also verify the tests_py36 folder
