+++ tox.ini
@@ -34,7 +34,7 @@ commands =
 basepython = python3.8
 deps = -e.[dev]
 commands =
-    flake8 gql tests
+    flake8 gql tests tests_py36
 
 [testenv:import-order]
 basepython=python3.8
